requirements.txt

fastapi
pydantic
uvicorn[standard]
azure-ai-projects
azure-identity
httpx
python-dotenv
semantic-kernel>=1.0.0
mcp>=1.1.0
